Deterministic hashing
Previously the ehrql variable qm node hashes were not deterministic for some situations e.g. the order of args in an "Or", or the order of items in a "case when". These are now sorted so the hashes shouldn't change too much making it easier to see genuine differences when the code is rerun